requirements.txt
django
djangorestframework
django-cors-headers
bcrypt
langchain-google-genai
google-generativeai
psycopg2-binary # Example for PostgreSQL
gunicorn
drf-yasg  # For Swagger/OpenAPI documentation
